Changing test duration from an int to a float

+++ AudioKit/Core Classes/AKOrchestra.m
@@ -51,7 +51,7 @@
     }
 }
 
-+ (void)testForDuration:(int)duration
++ (void)testForDuration:(float)duration
 {
     [[AKManager sharedManager] setIsLogging:YES];
     if (![[AKManager sharedManager] isRunning]) {
